Adds origin to grpc requests. This fixes a current issue with client/server communication failing. The AddOrigin service is added to Connections.

In order to test this change, a few changes were made:
* bind now takes an optional handle so that cores can be shared.
* run_until was added so that server can be shutdown gracefully.

+++ edgelet/edgelet-grpc/src/server/unix.rs
@@ -2,10 +2,12 @@
 
 #![cfg(unix)]
 use std::{fs, io};
+use std::fmt::Debug;
+use std::path::Path;
 
 use http::request::Request;
 use http::response::Response;
-use tokio_core::reactor::Core;
+use tokio_core::reactor::Handle;
 use tokio_io::IoStream;
 use tokio_uds::{UnixListener as UdsListener, UnixStream};
 use tower::NewService;
@@ -15,6 +17,7 @@ use url::Url;
 use server::{Listener, Server};
 
 /// A Listener that accepts connections over Unix Domain socket.
+#[derive(Debug)]
 pub struct UnixListener {
     inner: UdsListener,
 }
@@ -30,20 +33,26 @@ impl Listener for UnixListener {
     type Addr = ::std::os::unix::net::SocketAddr;
     type Connections = IoStream<(Self::Socket, Self::Addr)>;
 
-    fn bind<S, B>(address: Url, new_service: S) -> Result<Server<Self, S, B>, io::Error>
+    fn bind_handle<S, B>(
+        address: Url,
+        handle: &Handle,
+        new_service: S,
+    ) -> Result<Server<Self, S, B>, io::Error>
     where
-        S: NewService<Request = Request<RecvBody>, Response = Response<B>> + 'static,
+        S: NewService<Request = Request<RecvBody>, Response = Response<B>> + Debug + 'static,
+        S::InitError: Debug,
+        S::Error: Debug,
         B: Body + 'static,
     {
-        let core = Core::new()?;
-        let handle = core.handle();
+        if Path::new(address.path()).exists() {
+            fs::remove_file(address.path())?;
+        }
 
-        fs::remove_file(address.path())?;
         let listener = UdsListener::bind(address.path(), &handle).map(UnixListener::new)?;
 
         Ok(Server {
             new_service,
-            core,
+            handle: handle.clone(),
             listener,
         })
     }
